+++ project/jobs/tests/test_tasks.py
@@ -89,22 +89,12 @@ class RunScheduledJobsTest(BaseTest):
         Should block multiple existing runs of this task. That way, no job
         looped through in this task can get started in huey multiple times.
         """
-        with self.assertLogs(logger='jobs.utils', level='DEBUG') as cm:
-
-            # Mock a function called by the task, and make that function
-            # attempt to run the task recursively.
-            with mock.patch(
-                'jobs.tasks.get_scheduled_jobs', call_run_scheduled_jobs
-            ):
-                run_scheduled_jobs()
-
-        log_message = (
-            "DEBUG:jobs.utils:"
-            "Job [run_scheduled_jobs] is already pending or in progress."
-        )
-        self.assertIn(
-            log_message, cm.output,
-            "Should log the appropriate message")
+        # Mock a function called by the task, and make that function
+        # attempt to run the task recursively.
+        with mock.patch(
+            'jobs.tasks.get_scheduled_jobs', call_run_scheduled_jobs
+        ):
+            run_scheduled_jobs()
 
         self.assertEqual(
             Job.objects.filter(job_name='run_scheduled_jobs').count(), 1,